exit_interviews/requirements.txt
torch==2.1.0 
torchvision==0.16.0 
torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu121
tiktoken
setuptools-rust
openai-whisper